Moved some things and updated tests
