+++ source/GameInterface/GameInterface.csproj
@@ -296,22 +296,14 @@
     <Compile Include="Serialization\Dynamic\IDynamicModelService.cs" />
     <Compile Include="Serialization\Dynamic\IDynamicSerializer.cs" />
     <Compile Include="Serialization\ISerializer.cs" />
-    <Compile Include="Serialization\Models\CampaignTimeSurrogate.cs" />
-    <Compile Include="Serialization\Models\CharacterPerksSurrogate.cs" />
-    <Compile Include="Serialization\Models\CharacterSkillsSurrogate.cs" />
-    <Compile Include="Serialization\Models\CultureSurrogate.cs" />
-    <Compile Include="Serialization\Models\EquipmentElementSurrogate.cs" />
-    <Compile Include="Serialization\Models\EquipmentSurrogate.cs" />
-    <Compile Include="Serialization\Models\HeroDynamicSerializer.cs" />
-    <Compile Include="Serialization\Models\ItemComponentDynamicSerializer.cs" />
-    <Compile Include="Serialization\Models\ItemDynamicSerializer.cs" />
-    <Compile Include="Serialization\Models\ItemModifierSurrogate.cs" />
-    <Compile Include="Serialization\Models\KingdomSurrogate.cs" />
-    <Compile Include="Serialization\Models\MBObjectSurrogate.cs" />
-    <Compile Include="Serialization\DynamicModel\DynamicModelGenerator.cs" />
-    <Compile Include="Serialization\Models\MobilePartySurrogate.cs" />
-    <Compile Include="Serialization\Models\SettlementSurrogate.cs" />
-    <Compile Include="Serialization\Models\TextObjectSurrogate.cs" />
+    <Compile Include="Serialization\Surrogates\CampaignTimeSurrogate.cs" />
+    <Compile Include="Serialization\Surrogates\CharacterPerksSurrogate.cs" />
+    <Compile Include="Serialization\Surrogates\CharacterSkillsSurrogate.cs" />
+    <Compile Include="Serialization\Surrogates\CultureSurrogate.cs" />
+    <Compile Include="Serialization\Dynamic\Serializers\HeroDynamicSerializer.cs" />
+    <Compile Include="Serialization\Surrogates\KingdomSurrogate.cs" />
+    <Compile Include="Serialization\Surrogates\MobilePartySurrogate.cs" />
+    <Compile Include="Serialization\Surrogates\SettlementSurrogate.cs" />
     <Compile Include="Serialization\Surrogates\EquipmentElementSurrogate.cs" />
     <Compile Include="Serialization\Surrogates\EquipmentSurrogate.cs" />
     <Compile Include="Serialization\Dynamic\Serializers\ItemComponentDynamicSerializer.cs" />
