+++ source/GameInterface/Serialization/Dynamic/DynamicModelGenerator.cs
@@ -1,4 +1,5 @@
 ﻿using ProtoBuf.Meta;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
