--- source/GameInterface/Serialization/Models/CharacterPerksSurrogate.cs
@@ -1,7 +1,0 @@
-﻿namespace GameInterface.Serialization.Models
-{
-    public class CharacterPerksSurrogate
-    {
-        
-    }
-}
