+++ source/GameInterface/Serialization/Surrogates/CharacterPerksSurrogate.cs
@@ -1,0 +1,7 @@
+﻿namespace GameInterface.Serialization.Surrogates
+{
+    public class CharacterPerksSurrogate
+    {
+
+    }
+}
