+++ source/GameInterface/Serialization/Surrogates/KingdomSurrogate.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using ProtoBuf;
 using TaleWorlds.CampaignSystem;
 
-namespace GameInterface.Serialization.Models
+namespace GameInterface.Serialization.Surrogates
 {
     [ProtoContract]
     public class KingdomSurrogate
